CMake: Remove K64F dependency on MBED_TARGET_LABELS
Remove the dependency on the MBED_TARGET_LABELS generated from the
config system. Some drivers and the "platform" module are checking the
target labels to deduce what kind of target we're building for. This is
so they can conditionally add the correct source files to the build for
the current target.

The tools also turn the labels, features and component strings into
compile definitions. mbed-os relies on these definitions to
conditionally compile platform specific code. We can check if these
macros are set in CMake's INTERFACE_COMPILE_DEFINITIONS property instead
of querying the MBED_TARGET_LABELS list.

Breaking the dependency on the MBED_TARGET_LABELS list means we can move
all the target specific macro definitions to the target's CMakeList.txt.

The idea behind this commit is to move us toward an architecture where
we set the target compile definitions purely in CMake, rather than
fetching them from targets.json. This architecture would simplify the
Mbed config system significantly. mbed-tools would still need to
process the mbed_lib.json config and the parameters in the "config"
field in the targets.json objects. However, the pattern implemented
here removes the need for the tools to care about the labels,
extra_labels, components and features fields in targets.json. This
should simplify our configuration and build process.